pyproject.toml
[tool.poetry]
name = "tts-hw"
version = "0.1.0"
description = "DLA homework"
authors = ["Ivan Dmitriev <i.u.dmitriev@gmail.com>"]
readme = "README.md"

[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cu118"
priority = 'supplemental'

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
llvmlite = "^0.41.0"
numpy = "<1.26.0"
tqdm = "^4.66.1"
tensorboard = "^2.14.1"
matplotlib = "^3.8.0"
pandas = ">=1.0.0"
speechbrain = "^0.5.15"
datasets = "^2.14.5"
editdistance = "^0.6.2"
wandb = "^0.15.12"
pyctcdecode = "^0.5.0"
numba = "0.58.0"
torch-audiomentations = "^0.11.0"
torch = { version = "2.1.0", source="torch"}
torchaudio = { version = ">=2.0.0", source="torch"}
torchvision = { version = ">=0.12.0", source="torch"}
einops = "^0.7.0"
pytorch-beam-search = "^1.2.2"
flashlight-text = "^0.0.4"
kenlm = {url = "https://github.com/kpu/kenlm/archive/master.zip"}
gdown = "^4.7.1"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
pyworld = "^0.3.4"
waveglow = "^22.12.28"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
